+++ config/auth.js
@@ -1,5 +1,6 @@
 export default {
   hashRounds: 64,
-  saltSecret: process.env.AUTH_SALT || 'gdfg45667_%%^trterte',
+  saltSecret:
+    process.env.AUTH_SALT || console.error('AUTH_SALT is not defined'),
   isAuthWithVefificationFlow: true,
 };